+++ tests/e2e/game/full-game.test.js
@@ -88,6 +88,7 @@ describe("B - Full game of 31 players with all roles", () => {
                 additionalCards = game.additionalCards;
                 expect(game.options.repartition.isHidden).to.be.false;
                 expect(game.options.roles.areRevealedOnDeath).to.be.true;
+                expect(game.options.roles.seer.canSeeRoles).to.be.true;
                 done();
             });
     });